requirements.txt
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
docopt==0.6.2
idna==3.4
pipreqs==0.4.11
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
yarg==0.1.9
